Put the fetchWeatherData() call into the signage service. Added night icon as well.
